requirements.txt
pygame==1.9.6
open3d==0.9.0.0
tensorflow_gpu==2.2.0
transforms3d==0.3.1
keyboard==0.13.5
opencv_python==4.2.0.34
numpy==1.18.4
scipy
chumpy
pyvirtualcam